Create KeyboardToken instead of modifying Keyboard
