+++ app/src/components/index.tsx
@@ -26,6 +26,7 @@ export * from "./CopyToClipboardButton";
 export * from "./SectionHeading";
 export * from "./Empty";
 export * from "./exception";
+export * from "./KeyboardToken";
 
 // design system based components
 export * from "./alert";
